sharding: Package comments and responding to PR feedback
