+++ sharding/client/client.go
@@ -1,3 +1,6 @@
+// Package client provides an interface for interacting with a running ethereum full node.
+// As part of the initial phases of sharding, actors will need access to the sharding management
+// contract on the main PoW chain.
 package client
 
 import (
@@ -7,6 +10,7 @@ import (
 	"fmt"
 	"math/big"
 	"os"
+	"time"
 
 	ethereum "github.com/ethereum/go-ethereum"
 
@@ -17,8 +21,10 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/rpc"
+	"github.com/ethereum/go-ethereum/sharding"
 	"github.com/ethereum/go-ethereum/sharding/contracts"
 	cli "gopkg.in/urfave/cli.v1"
 )
@@ -27,7 +33,7 @@ const (
 	clientIdentifier = "geth" // Used to determine the ipc name.
 )
 
-// General Client for Collator. Communicates to Geth node via JSON RPC.
+// General client for Collator/Proposer. Communicates to Geth node via JSON RPC.
 
 type shardingClient struct {
 	endpoint  string             // Endpoint to JSON RPC
@@ -188,4 +194,46 @@ func dialRPC(endpoint string) (*rpc.Client, error) {
 		endpoint = node.DefaultIPCEndpoint(clientIdentifier)
 	}
 	return rpc.Dial(endpoint)
+}
+
+// initSMC initializes the sharding manager contract bindings.
+// If the SMC does not exist, it will be deployed.
+func initSMC(c *shardingClient) (*contracts.SMC, error) {
+	b, err := c.client.CodeAt(context.Background(), sharding.ShardingManagerAddress, nil)
+	if err != nil {
+		return nil, fmt.Errorf("unable to get contract code at %s: %v", sharding.ShardingManagerAddress, err)
+	}
+
+	// Deploy SMC for development only.
+	// TODO: Separate contract deployment from the sharding client. It would only need to be deployed
+	// once on the mainnet, so this code would not need to ship with the client.
+	if len(b) == 0 {
+		log.Info(fmt.Sprintf("No sharding manager contract found at %s. Deploying new contract.", sharding.ShardingManagerAddress.String()))
+
+		txOps, err := c.CreateTXOps(big.NewInt(0))
+		if err != nil {
+			return nil, fmt.Errorf("unable to intiate the transaction: %v", err)
+		}
+
+		addr, tx, contract, err := contracts.DeploySMC(txOps, c.client)
+		if err != nil {
+			return nil, fmt.Errorf("unable to deploy sharding manager contract: %v", err)
+		}
+
+		for pending := true; pending; _, pending, err = c.client.TransactionByHash(context.Background(), tx.Hash()) {
+			if err != nil {
+				return nil, fmt.Errorf("unable to get transaction by hash: %v", err)
+			}
+			time.Sleep(1 * time.Second)
+		}
+
+		log.Info(fmt.Sprintf("New contract deployed at %s", addr.String()))
+		return contract, nil
+	}
+
+	contract, err := contracts.NewSMC(sharding.ShardingManagerAddress, c.client)
+	if err != nil {
+
+	}
+	return contract, nil
 }
