+++ sharding/collator/collator.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
+	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/sharding"
 	"github.com/ethereum/go-ethereum/sharding/client"
 )
@@ -118,5 +119,24 @@ func submitCollation(shardID int64) error {
 	// them to finish up the collation. It will then need to broadcast the
 	// collation to the main chain using JSON-RPC.
 	log.Info("Submit collation function called")
+	return nil
+}
+
+// joinCollatorPool checks if the account is a collator in the SMC. If
+// the account is not in the set, it will deposit 100ETH into contract.
+func joinCollatorPool(c client.Client) error {
+
+	log.Info("Joining collator pool")
+	txOps, err := c.CreateTXOps(sharding.DepositSize)
+	if err != nil {
+		return fmt.Errorf("unable to intiate the deposit transaction: %v", err)
+	}
+
+	tx, err := c.SMCTransactor().Deposit(txOps)
+	if err != nil {
+		return fmt.Errorf("unable to deposit eth and become a collator: %v", err)
+	}
+	log.Info(fmt.Sprintf("Deposited %dETH into contract with transaction hash: %s", new(big.Int).Div(sharding.DepositSize, big.NewInt(params.Ether)), tx.Hash().String()))
+
 	return nil
 }
