+++ sharding/contracts/contracts.go
@@ -1,0 +1,5 @@
+// Package contracts consists of generated contracts for sharding.
+package contracts
+
+// DO NOT USE THIS FILE: It exists only for the package comment. Other files in this package are
+// automatically generated and cannot have a package comment.
